UI and UX edits
The previous version required the use of zipcode to track where the user wanted the weather. I have replaced this with city and country inputs instead of zipcode allowing more variety and simplicity because who actually remembers zipcodes? Also i've made it a tad prettier using some bootstrap and css3.