UX update
I copy pasted the dictoinary from the text file directly to the script and ran through it and removed the spaces in the keys and made all letters lower case. This way someone could type in a country name in whatever fashion and i can still retrieve the code associated with that country. However, im still at a loss as to how i can do this for the cities aswell.